requirements.txt
flask
flask-smorest
flask-cors
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-migrate
passlib
flask-jwt-extended